+++ src/main/java/in/oxidane/work/done/material/service/MaterialManufacturerService.java
@@ -1,5 +1,6 @@
 package in.oxidane.work.done.material.service;
 
+import in.oxidane.work.done.common.exception.ResourceNotFoundException;
 import in.oxidane.work.done.material.dto.MaterialManufacturerRequest;
 import in.oxidane.work.done.material.dto.MaterialManufacturerResponse;
 import java.util.List;
@@ -24,9 +25,9 @@ public interface MaterialManufacturerService {
      *
      * @param id The ID of the material manufacturer to retrieve
      * @return The MaterialManufacturerResponse if found
-     * @throws in.oxidane.work.done.exception.ResourceNotFoundException if manufacturer not found
+     * @throws ResourceNotFoundException if manufacturer not found
      */
-    MaterialManufacturerResponse getMaterialManufacturerById(int id);
+    MaterialManufacturerResponse getMaterialManufacturerById(Long id);
 
     /**
      * Get all material manufacturers
@@ -41,16 +42,16 @@ public interface MaterialManufacturerService {
      * @param id The ID of the material manufacturer to update
      * @param request The MaterialManufacturerRequest containing updated information
      * @return The updated MaterialManufacturerResponse
-     * @throws in.oxidane.work.done.exception.ResourceNotFoundException if manufacturer not found
+     * @throws ResourceNotFoundException if manufacturer not found
      * @throws RuntimeException if update fails
      */
-    MaterialManufacturerResponse updateMaterialManufacturer(int id, MaterialManufacturerRequest request);
+    MaterialManufacturerResponse updateMaterialManufacturer(Long id, MaterialManufacturerRequest request);
 
     /**
      * Delete material manufacturer by ID
      *
      * @param id The ID of the material manufacturer to delete
-     * @throws in.oxidane.work.done.exception.ResourceNotFoundException if manufacturer not found
+     * @throws ResourceNotFoundException if manufacturer not found
      */
-    void deleteMaterialManufacturer(int id);
+    void deleteMaterialManufacturer(Long id);
 }
