+++ src/main/java/in/oxidane/work/done/worker/repository/WorkerTypeRepository.java
@@ -5,6 +5,6 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface WorkerTypeRepository extends JpaRepository<WorkerType, Integer> {
+public interface WorkerTypeRepository extends JpaRepository<WorkerType, Long> {
     // Add custom query methods if needed
 }
